requirements/dev.txt
amqp==2.5.2
args==0.1.0
asgiref==3.4.1
attrs==21.2.0
Babel==2.8.0
backports.csv==1.0.7
bcrypt==3.2.0
beautifulsoup4==4.8.1
billiard==3.6.4.0
boto3==1.9.5
botocore==1.12.253
celery==4.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
clint==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==2.2.28
django-appconf==1.0.2
django-cachalot==2.3.0
django-celery-beat==1.5.0
django-celery-results==1.0.4
django-celerybeat-status==0.0.10
django-compressor==2.2
django-configurations==2.2
django-constance==2.3.0
django-cors-headers==2.5.2
django-extensions==2.1.2
django-extra-fields==2.0.2
django-extra-views==0.12.0
django-filter==2.1.0
django-health-check==3.7.0
django-js-asset==1.2.2
django-kaio==0.13.0
django-logentry-admin==1.0.4
django-migration-checker==0.4.3
django-minio-storage==0.3.1
django-modeltranslation==0.13.3
django-mptt==0.9.1
django-oauth-toolkit==1.5.0
django-ordered-model==3.1.1
django-picklefield==1.1.0
django-prometheus==1.0.15
django-redis==4.10.0
django-robots==3.1.0
django-rosetta==0.9.3
django-safedelete==0.5.1
django-solo==1.1.3
django-storages==1.7.1
django-tables2==2.0.6
django-timezone-field==3.1
django-yubin==1.5.0
djangorestframework==3.9.2
docutils==0.15.2
docx2txt==0.8
drf-chunked-upload==0.4.2
drf-renderer-xlsx==0.3.3
drf-yasg==1.13.0
elementpath==1.3.3
eradicate==0.2
et-xmlfile==1.1.0
flake8==3.6.0
flower==0.9.5
geographiclib==1.52
geopy==1.19.0
humanize==3.12.0
idna==3.3
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.0.1
jwcrypto==1.0
kafka-python==2.0.1
kombu==4.6.6
lockfile==0.12.2
Markdown==3.1
MarkupSafe==2.0.1
mccabe==0.6.1
minio==5.0.5
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
paramiko==2.8.0
Pillow==8.4.0
polib==1.1.1
prometheus-client==0.8.0
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pycparser==2.20
pyflakes==2.0.0
PyNaCl==1.4.0
PyPDF2==1.26.0
pyrsistent==0.18.0
pysftp==0.2.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2018.5
PyYAML==6.0
pyzmail36==1.0.4
raven==6.9.0
rcssmin==1.0.6
redis==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.0.12
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.1.13
shortuuid==0.5.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
social-auth-core==4.2.0
swagger-spec-validator==2.4.3
tablib==0.13.0
tornado==6.1
tweepy==3.8.0
uritemplate==4.1.1
urllib3==1.25.11
utm==0.4.2
uWSGI==2.0.18
uwsgi-tools==1.1.1
vine==1.3.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
xmlschema==1.0.16
zipp==3.6.0
